+++ src/main/java/uk/gov/hmcts/darts/audio/model/AudioBeingProcessedFromArchiveQueryResult.java
@@ -1,5 +1,5 @@
 package uk.gov.hmcts.darts.audio.model;
 
 public record AudioBeingProcessedFromArchiveQueryResult(Long mediaId,
-                                                        Integer armExternalObjectDirectoryId) {
+                                                        Long armExternalObjectDirectoryId) {
 }
